pyproject.toml
[tool.poetry]
name = "universal"
version = "0.1.0"
description = ""
authors = ["wwfyde <wwfyde@163.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.11"
fastapi = { extras = ["all"], version = "^0.109.2" }
requests = "^2.31.0"
discord-py = "^2.3.2"
loguru = "^0.7.2"
redis = "^5.0.1"
celery = {extras = ["redis"], version = "^5.3.6"}
sqlalchemy = "^2.0.28"
psycopg = {extras = ["binary", "pool"], version = "^3.1.18"}
mysql-connector-python = "^8.3.0"
pillow = "^10.2.0"
alembic = "^1.13.1"

[tool.poetry.group.dev.dependencies]
ipython = "^8.21.0"
pytest = "^8.0.1"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
